validation section

See merge request hoanglong2002.jvb/BE-AutoCareerBridge!70

+++ src/main/java/com/backend/autocarrerbridge/service/impl/SectionServiceImpl.java
@@ -60,8 +60,10 @@ public class SectionServiceImpl implements SectionService {
 
     @Override
     public SectionRequest deleteSection(int id) {
-        Section section =
-                sectionRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.SECTION_NOT_FOUND));
+        Section section = sectionRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.SECTION_NOT_FOUND));
+        if(!section.getMajors().isEmpty()){
+            throw new AppException(ErrorCode.SECTION_HAVE_MAJOR);
+        }
         sectionRepository.delete(section);
         return SectionConverter.convertToResponse(section);
     }
